kiosk_module/pyproject.toml
[build-system]
requires = ["hatchling>=1.26"]
build-backend = "hatchling.build"

[project]
name = "kiosk-module"
version = "1.0.0"
description = "GPS SMART KIOSK PCB serial communication and WebSocket bridge"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyserial>=3.5",
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0",
    "pynput>=1.7.7",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]
gui = [
    "PyQt5>=5.15",
]

[tool.hatch.build.targets.wheel]
packages = ["kiosk_module"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
